Move faction-specific spells out of SharedSpellSetup
