+++ src/WarcraftLegacies.Source/Factions/Sunfury/SunfurySpells.cs
@@ -1,4 +1,5 @@
 ﻿using MacroTools.Spells;
+using MacroTools.UnitTraits;
 using WarcraftLegacies.Source.Factions.Sunfury.Spells;
 using WarcraftLegacies.Source.Shared.Spells;
 
@@ -39,5 +40,20 @@ public static class SunfurySpells
       Interval = 0.1f
     };
     SpellRegistry.Register(siphoningRitual);
+
+    UnitTypeTraitRegistry.Register(new RestoreManaFromDamage(ABILITY_A11N_ARCANE_ABSORPTION_SUNFURY_STORMWIND)
+    {
+      ManaPerDamage = new LeveledAbilityField<float>
+      {
+        Base = 0.20f,
+        PerLevel = 0.20f
+      },
+      Effect = @"Abilities\Spells\Undead\ReplenishMana\SpiritTouchTarget.mdl"
+    }, new[]
+    {
+      UNIT_N0E7_BLOODWARDER_SUNFURY,
+      UNIT_H05Y_LORD_WIZARD_STORMWIND,
+      UNIT_N0E8_SKYSHIP_SUNFURY
+    });
   }
 }
